use singletorn for apple and google push providers
